+++ src/modules/db/sql/SqlWorkspace.tsx
@@ -14,6 +14,7 @@ import DatabaseStats from "../components/DatabaseStats";
 import type { StatsCache } from "../components/DatabaseStats";
 import TransferOverlay from "../components/TransferOverlay";
 import ErrorBanner from "../../../components/ErrorBanner";
+import DisconnectTab from "../components/DisconnectTab";
 import TunnelBanner from "../components/TunnelBanner";
 import { useWorkspaceError } from "../workspaceError";
 import Input from "../../../components/Input";
@@ -698,6 +699,7 @@ function SqlWorkspace({
               {t(labelKey)}
             </button>
           ))}
+          <DisconnectTab onDisconnect={onDisconnect} />
         </div>
       </div>